+++ packages/atomic/src/components/facets/facet-value/facet-value.tsx
@@ -1,44 +1,38 @@
-import {Component, Event, EventEmitter, h, Prop} from '@stencil/core';
+import {h} from '@stencil/core';
 import {randomID} from '../../../utils/utils';
 
-@Component({
-  tag: 'facet-value',
-  shadow: false,
-})
-export class FacetValue {
-  @Event() public facetValueSelected!: EventEmitter<void>;
-  @Prop()
-  public label!: string;
-  @Prop()
-  public numberOfResults!: number;
-  @Prop()
-  public isSelected!: boolean;
+export type FacetValueProps = {
+  facetValueSelected: VoidFunction;
+  label: string;
+  numberOfResults: number;
+  isSelected: boolean;
+};
 
-  public render() {
-    const id = randomID('input');
-    return (
-      <li
-        role="option"
-        class="flex flex-row items-center mt-2 cursor-pointer text-base"
+export const FacetValue = (props: FacetValueProps) => {
+  const id = randomID('input');
+  return (
+    <li
+      part="facet-value"
+      role="option"
+      class="flex flex-row items-center py-2 lg:py-1 cursor-pointer text-lg lg:text-base"
+    >
+      <input
+        type="checkbox"
+        checked={props.isSelected}
+        class="facet-value-checkbox w-5 h-5"
+        id={id}
+        name={id}
+        onClick={() => props.facetValueSelected()}
+      />
+      <label
+        htmlFor={id}
+        class="ml-3 flex flex-row text-on-background flex-grow cursor-pointer"
       >
-        <input
-          type="checkbox"
-          checked={this.isSelected}
-          class="facet-value-checkbox w-5 h-5"
-          id={id}
-          name={id}
-          onClick={() => this.facetValueSelected.emit()}
-        />
-        <label
-          htmlFor={id}
-          class="ml-3 flex flex-row text-on-background flex-grow cursor-pointer"
-        >
-          {this.label}
-          <span class="ml-auto self-end text-on-background-variant">
-            {this.numberOfResults}
-          </span>
-        </label>
-      </li>
-    );
-  }
-}
+        <span class="my-auto">{props.label}</span>
+        <span class="ml-auto my-auto self-end text-on-background-variant">
+          {props.numberOfResults}
+        </span>
+      </label>
+    </li>
+  );
+};