+++ packages/headless/src/api/analytics/analytics.ts
@@ -19,8 +19,9 @@ import {
   SearchHubSection,
   SearchSection,
 } from '../../state/state-sections';
-import {Context} from '../../features/context/context-state';
+import {ContextPayload} from '../../features/context/context-state';
 import {PreprocessRequest} from '../preprocess-request';
+import {getLanguage} from './shared-analytics';
 
 export type StateNeededByAnalyticsProvider = ConfigurationSection &
   Partial<
@@ -35,6 +36,10 @@ export type StateNeededByAnalyticsProvider = ConfigurationSection &
 export class AnalyticsProvider implements SearchPageClientProvider {
   constructor(private state: StateNeededByAnalyticsProvider) {}
 
+  public getLanguage() {
+    return getLanguage(this.state);
+  }
+
   public getSearchEventRequestPayload() {
     return {
       queryText: this.queryText,
@@ -48,7 +53,7 @@ export class AnalyticsProvider implements SearchPageClientProvider {
   public getBaseMetadata() {
     const {context} = this.state;
     const contextValues = context?.contextValues || {};
-    const formattedObject: Context = {};
+    const formattedObject: ContextPayload = {};
     for (const [key, value] of Object.entries(contextValues)) {
       const formattedKey = `context_${key}`;
       formattedObject[formattedKey] = value;