+++ packages/headless/src/controllers/relevance-inspector/headless-relevance-inspector.test.ts
@@ -28,6 +28,38 @@ describe('RelevanceInspector', () => {
     expect(buildRelevanceInspector(engine)).toBeTruthy();
   });
 
+  it('when enabling debug, should call the listener', () => {
+    const listenerSpy = jest.fn();
+    relevanceInspector.subscribe(listenerSpy);
+    relevanceInspector.enable();
+    expect(listenerSpy).toHaveBeenCalledTimes(1);
+  });
+
+  it('when enabling debug twice, should call the listener once', () => {
+    const listenerSpy = jest.fn();
+    relevanceInspector.subscribe(listenerSpy);
+    relevanceInspector.enable();
+    relevanceInspector.enable();
+    expect(listenerSpy).toHaveBeenCalledTimes(1);
+  });
+
+  it('when disabling, should call the listener', () => {
+    const listenerSpy = jest.fn();
+    relevanceInspector.enable();
+    relevanceInspector.subscribe(listenerSpy);
+    relevanceInspector.disable();
+    expect(listenerSpy).toHaveBeenCalledTimes(1);
+  });
+
+  it('when disabling twice, should call the listener once', () => {
+    const listenerSpy = jest.fn();
+    relevanceInspector.enable();
+    relevanceInspector.subscribe(listenerSpy);
+    relevanceInspector.disable();
+    relevanceInspector.disable();
+    expect(listenerSpy).toHaveBeenCalledTimes(1);
+  });
+
   it(`when initialized with enabled="true"
   it should dispatch an "enableDebug" action`, () => {
     initRelevanceInspector({initialState: {enabled: true}});