+++ packages/samples/headless-react/src/engine.ts
@@ -1,6 +1,15 @@
-import {HeadlessEngine, searchAppReducers} from '@coveo/headless';
+import {
+  HeadlessEngine,
+  searchAppReducers,
+  recommendationAppReducers,
+} from '@coveo/headless';
 
 export const engine = new HeadlessEngine({
   configuration: HeadlessEngine.getSampleConfiguration(),
   reducers: searchAppReducers,
+});
+
+export const recommendationEngine = new HeadlessEngine({
+  configuration: HeadlessEngine.getSampleConfiguration(),
+  reducers: recommendationAppReducers,
 });